// src/core/http/generated/models/admin-partners/updatePartnerRequest.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Partners
 * Filtered OpenAPI specification for Admin Partners.
 * OpenAPI spec version: 0.0.1
 */

export interface UpdatePartnerRequest {
  /** @maxLength 255 */
  name?: string;
  /**
     * @maxLength 255
     * @nullable
     */
  company_name?: string | null;
  /**
     * @maxLength 32
     * @nullable
     */
  document?: string | null;
  /**
     * @maxLength 255
     * @nullable
     */
  email?: string | null;
  /**
     * @maxLength 32
     * @nullable
     */
  phone?: string | null;
  /**
     * @maxLength 2048
     * @nullable
     */
  website?: string | null;
  /**
     * @maxLength 255
     * @nullable
     */
  instagram?: string | null;
  /** @nullable */
  notes?: string | null;
  is_active?: boolean;
}